refactor: refactoring the entire code
First, I performed Visual Studio's feature of calculating code metrics on the entire code.

During this process, I discovered that some metrics were overloaded and I made some corrections to them.

In the case of existing logic, the Form had all the logic on this system. But my suggested logic created a KeyHandler interface and an implementation of ChoenJiIn to separate core logic from the Form.

To achieve this, I added files are `IKeyHandler.cs`,` CheonJiInHandler.cs`.

One thing to worry about is that the constructor of the `CheonJiInKeyHandler` class must always include a method that executes the Form's task. This seems to generate serious a coupling of the Form class. So, I think that we must refactor this in the near future.

Except for these changes, the rest of the changes were employed by using Visual Studio's code cleanup profile #1.